Revert "AAAAAAAH!"
This reverts commit 2cf54de835a36cf16e2f40227b0687705a764a8c.

+++ core/lowl_sample.c
@@ -1,5 +1,4 @@
 #include "lowl_sample.h"
-#include <stdlib.h>
 
 /*********************************************************
  *                                                       *
@@ -52,10 +51,10 @@ size_t reservoirsampler_occupied(reservoirsampler* rs) {
 }
 
 void reservoirsampler_print(reservoirsampler* rs) {
-  printf("%u", (unsigned int) rs->sample[0]);
+  printf("%u", rs->sample[0]);
   size_t occupied = reservoirsampler_occupied(rs);
   for (size_t i = 1; i < occupied; ++i)
-    printf(" %u", (unsigned int) rs->sample[i]);
+    printf(" %u", rs->sample[i]);
   for (size_t i = occupied; i < rs->capacity; ++i)
     printf(" ()");
   printf("\n");
